he: Run most tests under global maintenance
Move test_set_global_maintenance to test_001, so that HE HA does not
interfere with other OST stuff. In particular, running OpenSCAP reports
on the hosts caused high CPU and memory load, which caused HA to stop
the engine VM.
